refactor: moved host waiting into classless comp
