+++ client/components/game/PlayerRemoved.js
