+++ client/components/quiz/PlayerFinished.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import {PieChart, Pie, Cell, Label} from 'recharts'
-import FinishedButtons from '../FinishedButtons'
+import FinishedButtons from '../game/FinishedButtons'
 
 const PlayerFinished = props => {
   const noData = !props.answerData[0].value && !props.answerData[1].value
