+++ client/components/quiz/PlayerView.js
@@ -5,8 +5,8 @@ import {setResponseThunk} from '../../store/user'
 import Welcome from '../Welcome'
 import {connect} from 'react-redux'
 import {Redirect} from 'react-router-dom'
-import PlayerDisconnected from './PlayerDisconnected'
-import PlayerRemoved from './PlayerRemoved'
+import PlayerDisconnected from '../game/PlayerDisconnected'
+import PlayerRemoved from '../game/PlayerRemoved'
 import PlayerFinished from './PlayerFinished'
 import PlayerChoices from './PlayerChoices'
